Move WebPushJobData to its own file
