+++ app/src/main/java/com/nextcloud/client/jobs/BackgroundJobManager.kt
@@ -12,6 +12,7 @@ import androidx.work.ListenableWorker
 import com.nextcloud.client.account.User
 import com.owncloud.android.datamodel.OCFile
 import com.owncloud.android.datamodel.SyncedFolder
+import com.owncloud.android.datamodel.WebPushJobData
 import com.owncloud.android.operations.DownloadType
 
 /**
@@ -132,38 +133,6 @@ interface BackgroundJobManager {
 
     fun startNotificationJob(subject: String, signature: String)
     fun startDecryptedNotificationJob(accountName: String, message: String)
-
-    sealed class WebPushJobData() {
-        abstract val inputData: Data
-
-        data class Register(
-            val accountName: String,
-            val url: String,
-            val uaPublicKey: String,
-            val auth: String
-        ): WebPushJobData() {
-            override val inputData = Data.Builder()
-                .putString(UnifiedPushWork.ACTION, UnifiedPushWork.ACTION_REGISTER)
-                .putString(UnifiedPushWork.EXTRA_ACCOUNT, accountName)
-                .putString(UnifiedPushWork.EXTRA_URL, url)
-                .putString(UnifiedPushWork.EXTRA_UA_PUBKEY, uaPublicKey)
-                .putString(UnifiedPushWork.EXTRA_AUTH, auth)
-                .build()
-        }
-        data class Activate(val accountName: String, val token: String): WebPushJobData() {
-            override val inputData = Data.Builder()
-                .putString(UnifiedPushWork.ACTION, UnifiedPushWork.ACTION_ACTIVATE)
-                .putString(UnifiedPushWork.EXTRA_ACCOUNT, accountName)
-                .putString(UnifiedPushWork.EXTRA_TOKEN, token)
-                .build()
-        }
-        data class Unregister(val accountName: String): WebPushJobData() {
-            override val inputData = Data.Builder()
-                .putString(UnifiedPushWork.ACTION, UnifiedPushWork.ACTION_UNREGISTER)
-                .putString(UnifiedPushWork.EXTRA_ACCOUNT, accountName)
-                .build()
-        }
-    }
 
     fun startWebPushJob(jobData: WebPushJobData)
